Added new options to encrypt a password.
The generate-password command can now be passed the password to generate.
This allows for the password to be known by the user. There are potential
issues with that mechanism (i.e. password appears in the command and thus
in ps -ef).

Updated the docs accordingly (and fixed a few issues too).